refactor cli a bit
